[feat]

+++ reporter/grpc.go
@@ -298,7 +298,7 @@ func (r *gRPCReporter) SendMetrics(metrics go2sky.RunTimeMetric) {
 		HeapAlloc:    metrics.HeapAlloc,
 		StackInUse:   metrics.StackInUse,
 		GcNum:        metrics.GcNum,
-		GcPauseTime:  metrics.GcPauseTime,
+		GcPauseTime:  float32(metrics.GcPauseTime),
 		GoroutineNum: metrics.GoroutineNum,
 		ThreadNum:    metrics.ThreadNum,
 		CpuUsedRate:  float32(metrics.CpuUsedRate),
